requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.13.0
stem==1.8.2
pysocks==1.7.1
lxml==4.9.3
urllib3==2.0.7
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0
aiohttp==3.9.0
asyncio==3.4.3
validators==0.22.0
